refactor: rename fetchers to preloaders
